+++ front/study_flow/lib/core/db_helper/users_db_helper.dart
@@ -1,15 +1,16 @@
+import 'package:flutter/foundation.dart';
 import 'package:study_flow/models/user_model.dart';
-import 'all_db_helper.dart'; // 위에서 만든 파일 import
+import 'all_db_helper.dart';
 
 class UsersDBHelper {
-  // 유저 추가
   static Future<int> insertUser(UserModel user) async {
+    if (kIsWeb) return 0;
     final db = await LocalDatabase.instance.database;
     return await db.insert('users', user.toMap());
   }
 
-  // 유저 조회
   static Future<UserModel?> selectUser() async {
+    if (kIsWeb) return null;
     final db = await LocalDatabase.instance.database;
     final result = await db.query(
       'users',
@@ -20,6 +21,7 @@ class UsersDBHelper {
   }
 
   static Future<List<UserModel>?> selectUsers() async {
+    if (kIsWeb) return null;
     final db = await LocalDatabase.instance.database;
     final result = await db.query('users');
     return result.isNotEmpty
@@ -27,25 +29,24 @@ class UsersDBHelper {
         : null;
   }
 
-  // 유저 수정
   static Future<int> updateUser(
     String id, {
     String? name,
     String? password,
     String? is_login,
   }) async {
+    if (kIsWeb) return 0;
     final db = await LocalDatabase.instance.database;
     final Map<String, dynamic> updates = {};
     if (name != null) updates['name'] = name;
     if (password != null) updates['password'] = password;
     if (is_login != null) updates['is_login'] = is_login;
-
     if (updates.isEmpty) return 0;
     return await db.update('users', updates, where: 'id = ?', whereArgs: [id]);
   }
 
-  // 유저 삭제
   static Future<int> deleteUser(String id) async {
+    if (kIsWeb) return 0;
     final db = await LocalDatabase.instance.database;
     return await db.delete('users', where: 'id = ?', whereArgs: [id]);
   }
